requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.0
azure-ai-openai==1.0.0
azure-search-documents==11.4.0
redis==5.0.1
pydub==0.25.1
PyYAML==6.0.1
websockets==12.0